refactoring: delete unused files and components
