+++ src/core/components/Footer.tsx
